+++ src/lib/PnP.Framework/Modernization/Cache/DictionaryIntConverter.cs
@@ -1,0 +1,55 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace PnP.Framework.Modernization.Cache
+{
+    internal class DictionaryIntConverter: JsonConverter<Dictionary<int, string>>
+    {
+        public override Dictionary<int, string> Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+        {
+            if (reader.TokenType != JsonTokenType.StartObject)
+            {
+                throw new JsonException();
+            }
+
+            var value = new Dictionary<int, string>();
+
+            while (reader.Read())
+            {
+                if (reader.TokenType == JsonTokenType.EndObject)
+                {
+                    return value;
+                }
+
+                string keyString = reader.GetString();
+
+                if (!int.TryParse(keyString, out int keyAsInt32))
+                {
+                    throw new JsonException($"Unable to convert \"{keyString}\" to System.Int32.");
+                }
+
+                reader.Read();
+
+                string itemValue = reader.GetString();
+
+                value.Add(keyAsInt32, itemValue);
+            }
+
+            throw new JsonException("Error Occured");
+        }
+
+        public override void Write(Utf8JsonWriter writer, Dictionary<int, string> value, JsonSerializerOptions options)
+        {
+            writer.WriteStartObject();
+
+            foreach (KeyValuePair<int, string> item in value)
+            {
+                writer.WriteString(item.Key.ToString(), item.Value);
+            }
+
+            writer.WriteEndObject();
+        }
+    }
+}